se creo el controller de compañia y cuentas

+++ Entities/Accounts/Activo.cs
@@ -3,16 +3,17 @@ using AriesWebApi.Entities.Interfaces;
 
 namespace AriesWebApi.Entities.Accounts
 {
+        
     public class Activo : ITipoCuenta
     {
         public TipoCuenta TipoCuenta { get { return TipoCuenta.Activo; } }
         public Comportamiento Comportamiento { get { return Comportamiento.Debito;  } }
 
-        public double SaldoActual(double saldo, double debito, double credito)
+        public decimal SaldoActual(decimal saldo, decimal debito, decimal credito)
         {
             return (saldo + debito - credito);
         }
-        public double SaldoMensual(double debito, double credito) {
+        public decimal SaldoMensual(decimal debito, decimal credito) {
             return (debito - credito);
         }
     }
