+++ Entities/Accounts/Ingreso.cs
@@ -7,7 +7,7 @@ namespace AriesWebApi.Entities.Accounts
     {
         public TipoCuenta TipoCuenta { get { return TipoCuenta.Ingreso; } }
         public Comportamiento Comportamiento { get { return Comportamiento.Credito; } }
-        public double SaldoActual(double saldo, double debito, double credito)
+        public decimal SaldoActual(decimal saldo, decimal debito, decimal credito)
         {
             return (saldo - debito + credito);
         }
@@ -17,7 +17,7 @@ namespace AriesWebApi.Entities.Accounts
         /// <param name="debito"></param>
         /// <param name="credito"></param>
         /// <returns></returns>
-        public double SaldoMensual(double debito, double credito)
+        public decimal SaldoMensual(decimal debito, decimal credito)
         {
                 /// credito - debito
             return (debito + credito);
